requirements-airflow.txt
pyathena
pymongo
minio
psycopg2-binary